+++ migrations/token/ERC20/polygon_bridging/fx-portal/FxERC20MintBurnRootTunnel@0.3_deploy.js
@@ -1,6 +1,6 @@
 const Contract_deploy = require('../../../../../src/templates/Contract/deploy');
 const {getNamedAccount, getContractAddress} = require('../../../../../src/helpers/templates');
-const {skipChainTypesExceptFor} = require('../../../../../src/helpers/common');
+const {multiSkip, skipNetworks, skipChainTypesExceptFor} = require('../../../../../src/helpers/common');
 
 module.exports = Contract_deploy('FxERC20MintBurnRootTunnel@0.3', {
   contract: 'FxERC20MintBurnRootTunnel',
@@ -13,6 +13,9 @@ module.exports = Contract_deploy('FxERC20MintBurnRootTunnel@0.3', {
   ],
   deterministicDeployment: true,
 });
-module.exports.skip = skipChainTypesExceptFor('ethereum');
+module.exports.skip = multiSkip(
+  skipNetworks(['sepolia']), // until fx-portal is supported on amoy
+  skipChainTypesExceptFor('ethereum'),
+);
 module.exports.tags = ['ERC20', 'PolygonFxPortal'];
 module.exports.dependencies = ['ForwarderRegistry@1.0_deploy'];